Ercot Magic Documentation

Installation

To install the package, use the following code:

using Pkg
Pkg.add("ErcotMagic")

Overview

This is the documentation for the Ercot Magic package. The Ercot API is a great tool for accessing data on the Texas electricity market. This package provides a simple interface for accessing data from the Ercot API. The package is designed to be easy to use and flexible, allowing you to access a wide range of data from the Ercot API.

Setup

You will need to obtain a token from the Ercot API to collect data. Go to the Ercot API and create a user name and password. Once you do that, you can use the get_auth_token function to obtain a token in the form required for pulling in data. This token will be used to authenticate your requests to the Ercot API. The token is valid for 60 minutes, so you will need to obtain a new token regularly, but this is handled automatically by the package.

To obtain a token, use the following code:

using ErcotMagic
token = get_auth_token()

Accessing Data

The API provides a number of endpoints that you can access. I have create a some endpoints as constants within this package for ease of operation, but you can use which ones matter for you.

The ERCOT API does not provide access to all public data as of 03/23/2024. You can access key data like DA LMP, RT LMP, and load/gen forecasts, but the 60 day report data is not available.

Without further ado, let's get some data. It's very simple to access the data, but you will need to know a bit about the parameters required for each query.

Getting Day-Ahead LMP Data

using ErcotMagic
da_prices = "https://api.ercot.com/api/public-reports/np4-190-cd/dam_stlmnt_pnt_prices?"
params = Dict("deliveryDateFrom" => "2024-02-01", "deliveryDateTo" => "2024-02-25")
da_dat = get_ercot_data(params, da_prices)

Endpoint URLS

You can also use a set of constants that are predefined for a few interesting endpoints. These are defined in the code, but here are a few examples:

Day Ahead Prices

da_prices = "https://api.ercot.com/api/public-reports/np4-190-cd/dam_stlmnt_pnt_prices?"

Real Time Prices

rt_prices = "https://api.ercot.com/api/public-reports/np6-970-cd/rtd_lmp_node_zone_hub?"

Hourly system-wide Mid-Term Load Forecasts (MTLFs) for all forecast models with an indicator for which forecast was in use by ERCOT at the time of publication for current day plus the next 7.

Ercot Load Forecast Endpoint

ercot_load_forecast = "https://api.ercot.com/api/public-reports/np3-566-cd/lf_by_model_study_area?"

Ercot Zone Load Forecast Endpoint

ercot_zone_load_forecast = "https://api.ercot.com/api/public-reports/np3-565-cd/lf_by_model_weather_zone?"

Solar System Forecast

solar_system_forecast = "https://api.ercot.com/api/public-reports/np4-737-cd/spp_hrly_avrg_actl_fcast?"

Wind System Forecast

wind_system_forecast = "https://api.ercot.com/api/public-reports/np4-732-cd/wpp_hrly_avrg_actl_fcast?"

Getting Real-Time LMP Data

rt_prices = "https://api.ercot.com/api/public-reports/np6-970-cd/rtd_lmp_node_zone_hub?"
params = Dict("RTDTimestampFrom" => "2024-02-01T00:00:00", 
                "RTDTimestampTo" => "2024-02-01T01:00:00",
                "settlementPoint" => "HB_NORTH")
rt_dat = get_ercot_data(params, rt_prices)

Getting Load Forecast Data

ercot_load_forecast = "https://api.ercot.com/api/public-reports/np3-566-cd/lf_by_model_study_area?"
params = Dict("deliveryDateFrom" => "2024-02-21", "deliveryDateTo" => "2024-02-25")
lf_dat = get_ercot_data(params, ercot_load_forecast)

Documentation of Functions

ErcotMagic.ercot_api_call — Method

Base call to ERCOT API

authorization => token_bearer

headers => Ocp-Apim-Subscription-Key => ENV["ERCOTKEY"]

url => https://api.ercot.com/api/public-reports/np3-966-er/60damenergyonlyoffers

sixtydamenergyonlyoffers = "https://api.ercot.com/api/public-reports/np3-966-er/60damenergyonlyoffers" response = ercotapicall(token["idtoken"], sixtydamenergyonly_offers)

ErcotMagic.ercot_api_url — Method

Function to formulate a url for ERCOT API based on params in kwargs

params = Dict("deliveryDateFrom" => "2021-08-01", "deliveryDateTo" => "2024-02-25") params2 = Dict("settlementPointName" => "HBNORTH") url = ercotapi_url(params)

try to cal

response = ercotapicall(token["id_token"], url)

DAM prices

dampricesurl = ercotapiurl(params, daprices) response = ercotapicall(token["idtoken"], dampricesurl)

RT prices

rtpricesurl = ercotapiurl(params, rtprices) response = ercotapicall(token["idtoken"], rtpricesurl)

Two day AS

twodayasurl = ercotapiurl(params, twodayAS) response = ercotapicall(token["id_token"], twodayasurl)

Sixty DAM awards

params = Dict("deliveryDateFrom" => "2021-08-01", "deliveryDateTo" => "2024-02-25") sixtydamawardsurl = ercotapiurl(params, sixtydamawards) response = ercotapicall(token["idtoken"], sixtydamawards_url)

ErcotMagic.get_auth_token — Method

A function to retreive the auth token

token = getauthtoken()

ErcotMagic.get_ercot_data — Method

Mega function to get the data from ERCOT API

Examples: params = Dict("deliveryDateFrom" => "2024-02-01", "deliveryDateTo" => "2024-02-25") dadat = getercotdata(params, daprices)

Real Time Prices for every five minutes

params = Dict("RTDTimestampFrom" => "2024-02-01T00:00:00", "RTDTimestampTo" => "2024-02-01T01:00:00", "settlementPoint" => "HBNORTH") rtdat = getercotdata(params, rt_prices)

Load Forecast

params = Dict("deliveryDateFrom" => "2024-02-01", "deliveryDateTo" => "2024-02-25") lfdat = getercotdata(params, ercotload_forecast)

Zone Load Forecast

params = Dict("deliveryDateFrom" => "2024-02-21", "deliveryDateTo" => "2024-02-25") lfdat = getercotdata(params, ercotzoneloadforecast)

Solar System Forecast

params = Dict("deliveryDateFrom" => "2024-02-21") lfdat = getercotdata(params, solarsystem_forecast)

Wind System Forecast

params = Dict("deliveryDateFrom" => "2024-03-21") lfdat = getercotdata(params, windsystem_forecast)

ErcotMagic.parse_ercot_response — Method

Takes a response object and returns a DataFrame

params = Dict("deliveryDateFrom" => "2024-02-01", "deliveryDateTo" => "2024-02-25")

dadat = parseercotresponse(ercotapicall(token["idtoken"], ercotapiurl(params, da_prices)))

#Note: RTD LMP includes all adders params = Dict("RTDTimestampFrom" => "2024-02-01T00:00:00", "RTDTimestampTo" => "2024-02-01T01:00:00") rtdat = parseercotresponse(ercotapicall(token["idtoken"], ercotapiurl(params, rt_prices)))